refactor : 입력 테스트 분리
